Keep the resolved url and use that later
When fetching a url, store the resolved url for later use. This is done, for example, when a user provides a url that redirects. We store the resulting url (after following the redirect), so that it doesn't need to follow the redirect each time we make a request. Thus, this change improves performance when dealing with redirects.